feat(app-crm-minimal): refactor
