+++ examples/app-crm-minimal/package.json
@@ -22,7 +22,6 @@
     "@refinedev/nestjs-query": "^1.0.0",
     "@refinedev/nextjs-router": "^5.3.2",
     "@uiw/react-md-editor": "^3.19.5",
-    "algoliasearch": "^4.19.1",
     "antd": "^5.0.5",
     "classnames": "^2.3.2",
     "cross-env": "^7.0.3",
@@ -32,9 +31,7 @@
     "next": "^13.5.4",
     "nookies": "^2.5.2",
     "react": "^18.0.0",
-    "react-dom": "^18.0.0",
-    "react-instantsearch": "^7.0.1",
-    "axios": "^0.26.1"
+    "react-dom": "^18.0.0"
   },
   "devDependencies": {
     "@graphql-codegen/cli": "^5.0.0",
