--- examples/app-crm-minimal/pages/companies/components/card-view/card/index.ts
@@ -1,2 +1,0 @@
-export { CompanyCard } from "./card";
-export { CompanyCardSkeleton } from "./skeleton";
