+++ examples/app-crm-minimal/pages/kanban/components/board/index.tsx
