+++ examples/app-crm-minimal/pages/kanban/components/button/add-stage-button.tsx
