+++ examples/app-crm-minimal/pages/kanban/components/button/index.ts
