+++ examples/app-crm-minimal/pages/kanban/components/column/index.module.css
