+++ examples/app-crm-minimal/pages/kanban/components/forms/index.ts
@@ -1,5 +1,3 @@
-export { CheckListForm } from "./checklist-form";
-export { CommentForm } from "./comment-form";
 export { DescriptionForm } from "./description-form";
 export { DueDateForm } from "./duedate-form";
 export { StageForm } from "./stage-form";
