+++ examples/app-crm-minimal/pages/kanban/components/headers/index.ts
