+++ examples/app-crm-minimal/pages/kanban/components/index.ts
@@ -1,8 +1,6 @@
 export { KanbanBoard, KanbanBoardSkeleton } from "./board";
 export { KanbanColumn, KanbanColumnSkeleton } from "./column";
 export {
-    CheckListForm,
-    CommentForm,
     DescriptionForm,
     DueDateForm,
     StageForm,
@@ -22,14 +20,8 @@ export {
     ProjectCardMemo,
     ProjectKanbanCardSkeleton,
 } from "./project-kanban-card";
-export { AccordionHeaderSkeleton } from "./accordion/accordion-header-skeleton";
-export { Accordion } from "./accordion/accordion";
 export { KanbanAddCardButton } from "./button/add-card-button";
 export { KanbanAddStageButton } from "./button/add-stage-button";
-export { CheckListInput } from "./checklist-input";
-export { CommentList } from "./comment-list";
 export { ProjectKanbanList } from "./project-kanban-list";
 export { ProjectModalEdit } from "./project-modal-edit";
 export { ProjectModalCreate } from "./project-modal-create";
-export { ProjectStageModalCreate } from "./project-stage-modal-create";
-export { ProjectStageModalEdit } from "./project-stage-modal-edit";
