+++ examples/app-crm-minimal/pages/kanban/components/item/index.tsx
