+++ examples/app-crm-minimal/pages/kanban/components/project-kanban-card/index.tsx
