+++ examples/app-crm-minimal/pages/kanban/index.page.tsx
@@ -1,4 +1,4 @@
-import { ProjectKanbanList } from "../components";
+import { ProjectKanbanList } from "./components";
 
 export default function ScrumboardKanbanListPage() {
     return <ProjectKanbanList />;
