--- examples/app-crm-minimal/pages/scrumboard/components/project-stage-modal-create/index.tsx
@@ -1,50 +1,0 @@
-import { useModalForm } from "@refinedev/antd";
-import { useInvalidate, useNavigation } from "@refinedev/core";
-
-import { Form, Input, Modal } from "antd";
-
-export const ProjectStageModalCreate = () => {
-    const invalidate = useInvalidate();
-    const { list } = useNavigation();
-    const { formProps, modalProps, close } = useModalForm({
-        action: "create",
-        defaultVisible: true,
-        resource: "taskStages",
-        meta: {
-            fields: ["id"],
-        },
-        onMutationSuccess: () => {
-            invalidate({ invalidates: ["list"], resource: "tasks" });
-        },
-        successNotification: () => {
-            return {
-                key: "create-stage",
-                type: "success",
-                message: "Successfully created stage",
-                description: "Successful",
-            };
-        },
-    });
-
-    return (
-        <Modal
-            {...modalProps}
-            onCancel={() => {
-                close();
-                list("tasks", "replace");
-            }}
-            title="Add new stage"
-            width={512}
-        >
-            <Form {...formProps} layout="vertical">
-                <Form.Item
-                    label="Title"
-                    name="title"
-                    rules={[{ required: true }]}
-                >
-                    <Input />
-                </Form.Item>
-            </Form>
-        </Modal>
-    );
-};
