--- examples/app-crm-minimal/pages/scrumboard/kanban/stages/create/index.page.tsx
@@ -1,12 +1,0 @@
-import {
-    ProjectKanbanList,
-    ProjectStageModalCreate,
-} from "pages/scrumboard/components";
-
-export default function ScrumboardKanbanStageCreatePage() {
-    return (
-        <ProjectKanbanList>
-            <ProjectStageModalCreate />
-        </ProjectKanbanList>
-    );
-}
