--- examples/app-crm-minimal/pages/scrumboard/kanban/stages/edit/[id].page.tsx
@@ -1,9 +1,0 @@
-import { ProjectKanbanList, ProjectStageModalEdit } from "../../../components";
-
-export default function ScrumboardKanbanStageEditPage() {
-    return (
-        <ProjectKanbanList>
-            <ProjectStageModalEdit />
-        </ProjectKanbanList>
-    );
-}
