+++ examples/app-crm-minimal/src/components/accordion/accordion-header-skeleton.tsx
