+++ examples/app-crm-minimal/src/components/accordion/accordion.tsx
