+++ examples/app-crm-minimal/src/components/accordion/index.ts
