+++ examples/app-crm-minimal/src/components/index.ts
@@ -12,3 +12,4 @@ export * from "./pagination-total";
 export * from "./avatar-group";
 export * from "./list-title-button";
 export * from "./markdown-editor";
+export * from "./accordion";
