+++ examples/app-crm-minimal/src/providers/data/fetch-wrapper.ts
@@ -1,4 +1,3 @@
-import { gql, request } from "@refinedev/nestjs-query";
 import nookies from "nookies";
 
 type Error = {
@@ -30,72 +29,10 @@ export const fetchWrapper = async (url: string, options: RequestInit) => {
     const error = getGraphQLErrors(body);
 
     if (error) {
-        if (shouldRefreshToken(error)) {
-            const tokens = await refreshTokens();
-            if (!tokens) {
-                throw error;
-            }
-
-            const newResponse = await customFetch(url, options);
-            return newResponse;
-        }
-
         throw error;
     }
 
     return response;
-};
-
-const refreshTokens = async () => {
-    const cookies = nookies.get();
-    const currentRefreshToken = cookies?.["refresh_token"];
-    if (!currentRefreshToken) return null;
-
-    try {
-        const response = await request<{
-            refreshToken: {
-                accessToken: string;
-                refreshToken: string;
-            };
-        }>(
-            "https://api.crm.refine.dev/graphql",
-            gql`
-                mutation refreshToken($refreshToken: String!) {
-                    refreshToken(refreshToken: $refreshToken) {
-                        accessToken
-                        refreshToken
-                    }
-                }
-            `,
-            {
-                refreshToken: currentRefreshToken,
-            },
-        );
-
-        nookies.set(null, "access_token", response.refreshToken.accessToken, {
-            maxAge: 3 * 24 * 60 * 60, // 3 days
-            path: "/",
-        });
-        nookies.set(null, "refresh_token", response.refreshToken.refreshToken, {
-            maxAge: 7 * 24 * 60 * 60, // 7 days
-            path: "/",
-        });
-
-        return response.refreshToken;
-    } catch (error) {
-        nookies.destroy(null, "access_token");
-        nookies.destroy(null, "refresh_token");
-        return null;
-    }
-};
-
-const shouldRefreshToken = (error: Error) => {
-    const cookies = nookies.get();
-    const refreshToken = cookies?.["refresh_token"];
-    if (!refreshToken) return false;
-
-    if (error.statusCode === "UNAUTHENTICATED") return true;
-    return false;
 };
 
 const getGraphQLErrors = (body: any): Error | null => {
