+++ examples/app-crm-minimal/src/providers/index.ts
@@ -1,3 +1,2 @@
 export * from "./auth";
 export * from "./data";
-export * from "./search-client";
